+++ src/main/java/edu/ie3/simona/api/exceptions/ConversionException.java
@@ -6,7 +6,7 @@
 
 package edu.ie3.simona.api.exceptions;
 
-public class ConversionException extends Exception {
+public class ConversionException extends RuntimeException {
 
   public ConversionException(final String message) {
     super(message);