+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Renaming power fields in `EvModel` [#208](https://github.com/ie3-institute/simonaAPI/issues/208)
-
+- Enhancing external data connections [#219](https://github.com/ie3-institute/simonaAPI/issues/219)
 
 ## [0.5.0] - 2024-08-09